Move proxy and echo middleware to /http
